Удалил функицю, перебрал немного схемы

+++ dictionary/tables/ServiceDetails.sql
@@ -1,7 +1,7 @@
 CREATE TABLE IF NOT EXISTS dictionary.servicedetails
 (
-    service_id BIGINT NOT NULL,
-    type_car   INT    NOT NULL,
+    service_id INT NOT NULL,
+    type_car   INT NOT NULL,
         CONSTRAINT uq_service_type_car UNIQUE (service_id, type_car),
     detail_id  BIGINT
 );
